mindsdb/requirements.txt
mindsdb_sdk
python-dotenv